requirements.txt
#"""
#Requirements file for Zazu Constitutional Intelligence System
#"""

# Core async framework
asyncio==3.4.3

# Database drivers
psycopg[binary]==3.1.18
redis==5.0.1

# Vector embeddings
sentence-transformers==2.3.1
# torch installed as dependency of sentence-transformers

# API framework (for future REST interfaces)
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging & observability
python-json-logger==2.0.7
rich==13.7.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
ruff==0.1.14
